# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: SnailMailSender
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "2.0"
servers:
  - url: http://localhost:8080
    description: Generated server url
tags:
  - name: SnailMailSender
    description: SnailMailSender
paths:
  /send/snailmail:
    post:
      tags:
        - SnailMailSender
      summary: Prepare snail mail for batch
      operationId: sendSnailMail
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendSnailMailRequest'
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Problem'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Problem'
        "200":
          description: Successful Operation
  /api-docs:
    get:
      tags:
        - API
      summary: OpenAPI
      operationId: getApiDocs
      responses:
        "200":
          description: OK
          content:
            application/yaml:
              schema:
                type: string
      x-auth-type: None
      x-throttling-tier: Unlimited
      x-wso2-mutual-ssl: Optional
components:
  schemas:
    Attachment:
      required:
        - content
        - name
      type: object
      properties:
        content:
          type: string
          description: The attachment (file) content as a BASE64-encoded string
          example: aGVsbG8gd29ybGQK
        name:
          type: string
          description: The attachment filename
          example: test.txt
        contentType:
          type: string
          description: The attachment content type
          example: application/pdf
          enum:
            - application/pdf
        envelopeType:
          type: string
          description: The envelope type for the letter
          example: WINDOWED
          enum:
            - WINDOWED
            - PLAIN
      description: Attachments
    SendSnailMailRequest:
      required:
        - batchId
        - department
      type: object
      properties:
        department:
          type: string
          description: Department and unit that should be billed
          example: "SBK(Gatuavdelningen, Trafiksektionen)"
        deviation:
          type: string
          description: If the letter to send deviates from the standard
          example: A3 Ritning
        batchId:
          type: string
          description: Batch id to be used for creating a csv-file
          example: 6a5c3d04-412d-11ec-973a-0242ac130043
        partyId:
          type: string
          description: Party id for the person the letter should be sent to
          example: 6a5c3d04-412d-11ec-973a-0242ac130003
        attachments:
          type: array
          description: Attachments
          items:
            $ref: '#/components/schemas/Attachment'
    Problem:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: '#/components/schemas/StatusType'
    StatusType:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        reasonPhrase:
          type: string
  securitySchemes: {}